refactor fluorocaron model
